fix(config): Remove input file as required config option
The whole idea behind mandatory input file specification revolves around
the fact that some banks may not provide a consistent file name to their
exports, so the tool needs to know what file to read. This makes sense
in the context of the CLI, but not in the context of the config file, as
you may not be using a custom, static naming scheme for saving the file,
in which case, you can't set this property, and must always specify it
using the CLI flags instead.